*** install() is broken ***
        B:\code\apt>apt install iconv
        iconv
        error: i not in [curr]
        error: i not installed

Looks like the so called "not useful" missing() I messed with a bit ago is
important after all, and leads to the first real down-the-rabbit-hole mess
encountered in this round of refactoring. Reminding me that install() REALLY
needs some careful attention and documenting so it's easier to understand
what the heck it's doing.