setup: use parameter for target folder (a big change)
Initial working process to create the an empty skeleton folder tree,
install.db and setup config files. The calling is a little messy,
because it has to be initialized before the command line is parsed (and
thus we don't know all the standard config values yet).

It works, but needs tightening up.