INFO - show name, ver, etc. for PKG
This commit is overloaded, should have been split into several. A new
top-level command is introduced, 'info'. It relies on get_info() which
builds a dict of useful pkg attributes. I totally forgot I already
wrote this! Reporting function was pulled out of get_info() to make the
command - just smarter separation of fucntions.

Another new thing is having the command itself show help-for-cmd when
parameters are empty. This works, but we broke top level usage and help
in process.