comment out do_download info-izing
...it just made things worse, broke download, md5, and messed with
install. They should all work again, but there are still problems with
install at least (it's too silent when the package is already
installed). I still believe in the overall info-izing goal, but need to
proceed more thoughtfully.

One thing that has come to light is that we need more atoms in
get_info(), along with meaningful and intuitive names. Such as
"cache_dir", "server_root", "release_root", and so on. Strictly
speaking, these probably shouldn't be in the pkg dict, but in a parent
object -- basically the role globals are in now. I need to draw a
visual map. The relationships and occurences are too much for me to see
in my head, and text doesn't convery meaning rapidly enough.