return dict of Nones when rc not exists
A bit ambivalent about this, since we're hardcoding the key names into
the function, but I like it better than using guard statements on every
caller.

Also note new pattern template for DEBUG. If true print noisy stuff,
otherwise be quiet. I wanted to do use just `if debug:` instead of the
more verbose `if debug == True:`, but because there is that debug()
func I've not yet figured out that doesn't work. TODO: figure the func
out, remove it or recast into something useful for me.