#18: write setuprc to disk, brokenish
Func exists, but isn't being triggered with `apt update`, but is on `apt
download ...`. At moment it writes to a test file, and this works. What
I'm not sure about is if it's okay to write "None" to the rc file, or
if it should just be silent if it doesn't have a real value. Todo: ask
@jef-n.