clean up debug reporting without --debug flag (install still broken)
At present we're either installing the primary package and missing the
dependencies, or vice versa. The function is getting way to long with
lots of duplicate code. Time to trim and think things through more
carefully. Clearly the list clean stuff can be shoved into a function
and reused many times.

I'm done for tonight though.