work around for local file alternate ini, #41
If any of local indicators are seen in the 1st 2 characters of the
option string then we bypass requests module. It's a bit simplistic but
I think robust enough for us; we only use requests/urllib to write a
copy of the file and then work from that locally anway.

Leaving the ticket open for a bit because this should be a function (so we don't create yet another global for the pile).