setup_ini_from(): fix \\server\share matching
All patterns in the docstring have been tested and work on 2 of my
machines.

Strictly speaking, this function is generic and not specific to
setup_ini. If `path-string` is local, it merely returns the string
back, unaltered. When `path-string` is a url it downloads the contents
at the url to a local temp file and returns path to that instead.