Makes cache option work, close #47
Works, but wordy. Strictly speaking only the cache_dir IF needs
`globals()` phrasing, because it's the only variable potentially being
assigned inside the function.

Thank you @orokusaki for the cleanest explanation of global/local
variable interaction I've encountered. (Or the the right explanation at
the right time for me anyway!)
http://stackoverflow.com/a/10852003/14420