close "Read the Docs apt module page is empty", #48
RTD requires modules relying on Windows-only DLLs etc. be "mocked". It
doesn't need to be able to actually run the code, just process for
docstrings etc.

Initially I thought it was ctypes itself that needed to be mocked, but
it turns out I needed to work closer to home and _mock the module which
calls ctypes_, `knownpaths.py`.

Thank you to @Dunes on Stack Overflow, whose comment I thought was
off-track and not going to help. However it gave just enough of a
turning to my mind and path of investigation to land me in the right
place after all. Not all teachings look like teaching!

http://stackoverflow.com/questions/30950870/mock-import-of-ctypes-fails-with-unsupported-operand-on-read-the-docs