Remove --arch flag, leaving only --bits
Not enough reason to have both, lets keep it simple. The only argument I
can come up with is that 'arch' is the more correct term, and 'x86_64'
is what actually governs things internally. However this is an
implementation detail and users shouldn't have to contend with it;
requires more typing too.

--HG--
branch : bit-ness